moved the animation functions in a separate module
